requirements.txt
django~=5.0.2
djangorestframework~=3.14.0
django-simple-history~=3.4.0
psycopg2